refactor(classes): rename class files to index.ts
